requirements.txt
fastapi
uvicorn[standard]
pydantic
tensorflow
scikit-learn
pandas
psycopg2-binary
datasets
numpy